requirements.txt
openai
dotenv
requests
django-webpush